feature: ADD DBfactory DatabaseBrickas and MySQL database.
