+++ src/adapter/database.py
@@ -1,4 +1,5 @@
 from src.infra.database.dynamodb import Dynamodb
+from src.infra.database.mysql import MySQL
 
 
 class DatabaseFactory:
@@ -6,5 +7,12 @@ class DatabaseFactory:
     def create_database_gpt(db_type: str = "dynamodb"):
         if db_type == "dynamodb":
             return Dynamodb()
+        else:
+            raise ValueError(f"Unknown database type: {db_type}")
+
+    @staticmethod
+    def create_database_brickas(db_type: str = "mysql"):
+        if db_type == "mysql":
+            return MySQL()
         else:
             raise ValueError(f"Unknown database type: {db_type}")
